Remove analysis panel and analyzer module (#147)
The on-demand analysis panel (A/I keys) is superseded by Ask Betty,
which covers the same functionality more flexibly without consuming
40 columns of screen real estate. Extract GoalExtractor into a new
goals.py module (still used by the Agent), delete analyzer.py, and
clean all references from store, TUI, config, CLI, export, and models.
