+++ src/betty/models.py
@@ -6,7 +6,6 @@ from enum import Enum
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from .analyzer import Analysis
     from .github import PRInfo
 
 
@@ -47,7 +46,6 @@ class Turn:
     critic_sentiment: str | None = None  # "progress" | "concern" | "critical"
     task_operation: tuple[str, dict[str, Any]] | None = None  # (operation_type, data) for task tools
     annotation: str | None = None  # User-provided annotation
-    analysis: "Analysis | None" = None  # On-demand LLM analysis
 
 
 @dataclass
