+++ src/betty/store.py
@@ -10,14 +10,6 @@ from .agent import Agent
 from .agent_models import SessionReport
 from .alerts import Alert, AlertLevel, check_turn_for_alerts, send_system_notification
 from .github import PRDetector
-from .analyzer import (
-    ANALYZER_SYSTEM_PROMPT,
-    RANGE_SYSTEM_PROMPT,
-    Analysis,
-    Analyzer,
-    make_analysis_cache_key,
-    make_range_cache_key,
-)
 from .cache import AnnotationCache, SummaryCache
 from .config import load_config
 from .models import Session, TaskState, Turn
@@ -52,7 +44,7 @@ class EventStore:
         self._branch_cache: dict[str, str | None] = {}  # project_dir -> branch name
         self._pr_detector = PRDetector()
 
-        # Load config and initialize summarizer + analyzer + agent
+        # Load config and initialize summarizer + agent
         config = load_config()
         self._agent: Agent | None = Agent(config.agent, llm_config=config.llm) if config.agent.enabled else None
         self._summarizer = Summarizer(
@@ -62,15 +54,8 @@ class EventStore:
             summary_style=config.summary.style,
             custom_summary_prompt=config.summary.custom_prompt,
         )
-        self._analyzer = Analyzer(
-            model=config.llm.model,
-            api_base=config.llm.api_base,
-            api_key=config.llm.api_key,
-            analyzer_config=config.analyzer,
-        )
         self._summary_cache = SummaryCache()  # single cache for both assistant and tool summaries
         self._annotation_cache = AnnotationCache()  # user annotations
-        self._range_analyses: dict[tuple[int, int], Analysis] = {}  # (start_turn_num, end_turn_num) -> Analysis
 
     def start_watching(
         self,
@@ -790,201 +775,6 @@ class EventStore:
                     return True
             return False
 
-    def analyze_turn(self, turn: Turn) -> bool:
-        """Submit a turn for on-demand LLM analysis.
-
-        Args:
-            turn: The turn to analyze
-
-        Returns:
-            True if analysis was submitted, False if already analyzed or loaded from cache
-        """
-        # Already analyzed successfully (summary doesn't start with "[")
-        if turn.analysis and not turn.analysis.summary.startswith("["):
-            return False
-
-        # Clear error analysis for retry
-        if turn.analysis and turn.analysis.summary.startswith("["):
-            turn.analysis = None
-
-        # Get active session
-        with self._lock:
-            if not self._active_session_id:
-                return False
-            session = self._sessions.get(self._active_session_id)
-            if not session:
-                return False
-
-        # Build context and check cache
-        context = self._analyzer._context_manager.build_context(session, turn)
-        cache_key = make_analysis_cache_key(
-            turn, context, self._analyzer.model, ANALYZER_SYSTEM_PROMPT
-        )
-
-        cached = self._summary_cache.get(cache_key)
-        if cached:
-            import json
-            try:
-                data = json.loads(cached)
-                turn.analysis = Analysis(
-                    summary=data["summary"],
-                    critique=data["critique"],
-                    sentiment=data["sentiment"],
-                    word_count=data["word_count"],
-                    context_word_count=data["context_word_count"],
-                    local_goal=data.get("local_goal"),
-                )
-                # Notify listeners so TUI refreshes
-                for listener in self._turn_listeners:
-                    try:
-                        listener(turn)
-                    except Exception:
-                        pass
-                return False
-            except (json.JSONDecodeError, KeyError):
-                pass  # Fall through to re-analyze
-
-        # Submit for analysis
-        self._analyzer.analyze_async(
-            session, turn, self._make_analysis_callback(turn, cache_key)
-        )
-        return True
-
-    def get_range_analysis(self, range_key: tuple[int, int]) -> Analysis | None:
-        """Get cached range analysis result."""
-        return self._range_analyses.get(range_key)
-
-    def clear_range_analyses(self) -> None:
-        """Clear all stored range analyses."""
-        self._range_analyses.clear()
-
-    def analyze_range(self, turns: list[Turn]) -> bool:
-        """Submit a range of turns for analysis.
-
-        Results are stored internally and accessible via get_range_analysis().
-        Turn listeners are notified when analysis completes (same as single-turn).
-
-        Args:
-            turns: The turns in the range
-
-        Returns:
-            True if submitted for analysis, False if already analyzed or loaded from cache
-        """
-        if not turns:
-            return False
-
-        range_key = (turns[0].turn_number, turns[-1].turn_number)
-
-        # Already analyzed successfully
-        existing = self._range_analyses.get(range_key)
-        if existing and not existing.summary.startswith("["):
-            return False
-
-        # Clear error analysis for retry
-        if existing and existing.summary.startswith("["):
-            del self._range_analyses[range_key]
-
-        # Get active session
-        with self._lock:
-            if not self._active_session_id:
-                return False
-            session = self._sessions.get(self._active_session_id)
-            if not session:
-                return False
-
-        # Build context and check cache
-        context = self._analyzer._context_manager.build_context_for_range(
-            session, turns, span_analyses=self._range_analyses
-        )
-        cache_key = make_range_cache_key(
-            turns, context, self._analyzer.model, RANGE_SYSTEM_PROMPT
-        )
-
-        cached = self._summary_cache.get(cache_key)
-        if cached:
-            import json
-            try:
-                data = json.loads(cached)
-                analysis = Analysis(
-                    summary=data["summary"],
-                    critique=data["critique"],
-                    sentiment=data["sentiment"],
-                    word_count=data["word_count"],
-                    context_word_count=data["context_word_count"],
-                    local_goal=data.get("local_goal"),
-                )
-                self._range_analyses[range_key] = analysis
-                # Notify turn listeners so TUI refreshes
-                for listener in self._turn_listeners:
-                    try:
-                        listener(turns[0])
-                    except Exception:
-                        pass
-                return False
-            except (json.JSONDecodeError, KeyError):
-                pass  # Fall through to re-analyze
-
-        # Submit for analysis — pass existing span analyses for nested summarization
-        self._analyzer.analyze_range_async(
-            session, turns, self._make_range_analysis_callback(turns, range_key, cache_key),
-            span_analyses=self._range_analyses,
-        )
-        return True
-
-    def _make_range_analysis_callback(
-        self, turns: list[Turn], range_key: tuple[int, int], cache_key: str
-    ) -> "Callable[[Analysis, bool], None]":
-        """Create callback that stores range analysis and notifies listeners."""
-        def callback(analysis: Analysis, success: bool) -> None:
-            self._range_analyses[range_key] = analysis
-            if success:
-                import json
-                self._summary_cache.set(
-                    cache_key,
-                    json.dumps({
-                        "summary": analysis.summary,
-                        "critique": analysis.critique,
-                        "sentiment": analysis.sentiment,
-                        "word_count": analysis.word_count,
-                        "context_word_count": analysis.context_word_count,
-                        "local_goal": analysis.local_goal,
-                    }),
-                )
-            # Notify turn listeners so TUI refreshes (same pattern as single-turn)
-            for listener in self._turn_listeners:
-                try:
-                    listener(turns[0])
-                except Exception:
-                    pass
-        return callback
-
-    def _make_analysis_callback(
-        self, turn: Turn, cache_key: str
-    ) -> "Callable[[Analysis, bool], None]":
-        """Create callback that updates turn analysis and notifies listeners."""
-        def callback(analysis: Analysis, success: bool) -> None:
-            turn.analysis = analysis
-            if success:
-                import json
-                self._summary_cache.set(
-                    cache_key,
-                    json.dumps({
-                        "summary": analysis.summary,
-                        "critique": analysis.critique,
-                        "sentiment": analysis.sentiment,
-                        "word_count": analysis.word_count,
-                        "context_word_count": analysis.context_word_count,
-                        "local_goal": analysis.local_goal,
-                    }),
-                )
-            # Notify turn listeners to refresh TUI
-            for listener in self._turn_listeners:
-                try:
-                    listener(turn)
-                except Exception:
-                    pass
-        return callback
-
     def load_annotations_for_session(self, session_id: str) -> None:
         """Load cached annotations for a session's turns."""
         with self._lock:
@@ -1011,7 +801,6 @@ class EventStore:
             watcher.stop()
 
         self._summarizer.shutdown()
-        self._analyzer.shutdown()
         if self._agent:
             self._agent.shutdown()
